Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.9",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "opencv-python>=4.10.0.84",
    "numpy>=2.1.0",
    "pillow>=11.0.0",
    "structlog>=24.4.0",
    "ultralytics>=8.3.50",
    "requests>=2.32.3",
]
